Fix: renamed 'Presistence' to 'Persistence' and updated namespaces
